level up testing

+++ contracts/Quest.sol
@@ -11,7 +11,7 @@ contract Quest is Ownable {
   ///@dev Quests with recommendation of stats (example 20 of strength) if below 20 we remove the lack to the % of chance of success from 0 to 100
   ///@dev If no more difference in success
   struct QuestStruct {
-    uint8 exp;
+    uint16 exp;
     uint8 percentDifficulty; //0 - 100
     uint8[] stats; //required
     uint8[] items;
@@ -39,7 +39,7 @@ contract Quest is Ownable {
   function setQuest(
     uint256 id,
     uint256 time,
-    uint8 exp,
+    uint16 exp,
     uint8 percentDifficulty,
     uint8[] memory stats,
     uint8[] memory items
